vote_state_mismatch: timestamps need to be signed and not unsigned

+++ src/flamenco/repair/fd_repair.c
@@ -470,7 +470,7 @@ fd_repair_send_requests( fd_repair_t * glob ) {
       fd_repair_window_index_t * wi = &protocol.inner.window_index;
       fd_hash_copy(&wi->header.sender, glob->public_key);
       fd_hash_copy(&wi->header.recipient, &active->key);
-      wi->header.timestamp = (ulong)glob->now/1000000LU;
+      wi->header.timestamp = glob->now/1000000L;
       wi->header.nonce = n;
       wi->slot = ele->dupkey.slot;
       wi->shred_index = ele->dupkey.shred_index;
@@ -482,7 +482,7 @@ fd_repair_send_requests( fd_repair_t * glob ) {
       fd_repair_highest_window_index_t * wi = &protocol.inner.highest_window_index;
       fd_hash_copy(&wi->header.sender, glob->public_key);
       fd_hash_copy(&wi->header.recipient, &active->key);
-      wi->header.timestamp = (ulong)glob->now/1000000LU;
+      wi->header.timestamp = glob->now/1000000L;
       wi->header.nonce = n;
       wi->slot = ele->dupkey.slot;
       wi->shred_index = ele->dupkey.shred_index;
@@ -494,7 +494,7 @@ fd_repair_send_requests( fd_repair_t * glob ) {
       fd_repair_orphan_t * wi = &protocol.inner.orphan;
       fd_hash_copy(&wi->header.sender, glob->public_key);
       fd_hash_copy(&wi->header.recipient, &active->key);
-      wi->header.timestamp = (ulong)glob->now/1000000LU;
+      wi->header.timestamp = glob->now/1000000L;
       wi->header.nonce = n;
       wi->slot = ele->dupkey.slot;
       break;
