+++ src/flamenco/runtime/fd_blockstore.c
@@ -542,7 +542,7 @@ fd_blockstore_publish( fd_blockstore_t * blockstore, ulong new_root_slot ) {
   ulong pruned_slot_cnt  = 0UL;
   ulong scanned_slot_cnt = 0UL;
   fd_wksp_t * wksp       = fd_blockstore_wksp( blockstore );
-  
+
   /* If root is */
   if( blockstore->root==0 ) {
     blockstore->root = new_root_slot;
@@ -563,14 +563,14 @@ fd_blockstore_publish( fd_blockstore_t * blockstore, ulong new_root_slot ) {
   ulong * prune_deque = fd_wksp_laddr_fast( wksp, blockstore->slot_prune_deque_gaddr );
 
   fd_blockstore_slot_prune_deque_push_tail( prune_deque, blockstore->root );
-  
+
   while( !fd_blockstore_slot_prune_deque_empty( prune_deque ) ) {
     ulong curr = fd_blockstore_slot_prune_deque_pop_head( prune_deque );
     scanned_slot_cnt++;
-    
+
     ulong * next_slot     = NULL;
     ulong   next_slot_len = 0UL;
-    
+
     int err = fd_blockstore_next_slot_query( blockstore, curr, &next_slot, &next_slot_len );
     if( FD_UNLIKELY( err!=FD_BLOCKSTORE_OK ) ) {
       fd_blockstore_slot_prune_deque_remove_all( prune_deque );
@@ -593,7 +593,7 @@ fd_blockstore_publish( fd_blockstore_t * blockstore, ulong new_root_slot ) {
 
   prune_time_ns += fd_log_wallclock();
 
-  FD_LOG_NOTICE(( "blockstore publish - new_root_slot: %lu, prev_root_slot: %lu, slots pruned: %lu, slots scanned: %lu, elapsed: %6.6f ms", 
+  FD_LOG_NOTICE(( "blockstore publish - new_root_slot: %lu, prev_root_slot: %lu, slots pruned: %lu, slots scanned: %lu, elapsed: %6.6f ms",
       new_root_slot, blockstore->root,
       pruned_slot_cnt, scanned_slot_cnt, (double)prune_time_ns*1e-6));
 
@@ -789,9 +789,9 @@ fd_buf_shred_insert( fd_blockstore_t * blockstore, fd_shred_t const * shred ) {
     fd_memset( slot_meta, 0, sizeof( fd_slot_meta_t ) );
 
     /* the "reference tick" is the tick at the time the leader prepared the entry batch */
-    ulong reference_tick             = shred->data.flags & FD_SHRED_DATA_REF_TICK_MASK;
-    ulong ms                         = reference_tick * FD_MS_PER_TICK;
-    ulong now                        = (ulong)fd_log_wallclock() / 1000000UL;
+    long reference_tick              = shred->data.flags & FD_SHRED_DATA_REF_TICK_MASK;
+    long ms                          = reference_tick * FD_MS_PER_TICK;
+    long now                         = fd_log_wallclock() / 1000000L;
     slot_meta->slot                  = slot_entry->slot;
     slot_meta->consumed              = ULONG_MAX;
     slot_meta->received              = 0;
